+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.4.0 (unreleased)
+
+- Add `DefinitionType` and `BindableType` bases on top of `BaseType`.
+- Add `CollectionType` utility type for gather types into single type.
+
+
 ## 0.3.0 (2022-04-25)
 
 - Fix "dependency required" error raised for GraphQL `Float` scalar type.